Independent logic for getting image
