+++ apps/Imagine.js
@@ -1,6 +1,6 @@
 import plugin from '../../../lib/plugins/plugin.js'
-import Config from '../components/config/config.js'
 import axios from 'axios'
+import { getPic } from '../components/midjourney/getPic.js'
 import { imagine } from '../components/midjourney/upImagine.js'
 import { parseImg } from '../utils/utils.js'
 import { getResults } from '../utils/task.js'
@@ -57,17 +57,7 @@ export class Imagine extends plugin {
 				e.reply(`生成图像失败，请查看控制台输出`)
 				return true
 			} else {
-				let configs = Config.getSetting()
-				const img = await axios.get(task.imageUrl,
-					{
-						responseType: 'arraybuffer',
-						proxy: configs.proxy.host && configs.proxy.port ? {
-							protocol: 'http',
-							host: `${configs.proxy.host}`, port: `${Number(configs.proxy.port)}`
-						} : undefined
-					}
-				)
-				let base64 = Buffer.from(img.data, 'binary').toString('base64');
+				const base64 = await getPic(task)
 				let resReply = await e.reply([{ ...segment.image(`base64://${base64}`), origin: true }, `任务耗时：${(task.finishTime - task.startTime) / 1000}s`], true)
 				if (!resReply) {
 					e.reply(`发送图像失败，可能是因为图像过大，或无法访问图像链接\n图像链接：${task.imageUrl}`)
